default addresses can no longer be deleted

+++ src/app/Http/Controllers/AddressesController.php
@@ -11,6 +11,7 @@ use LaravelEnso\Core\app\Exceptions\EnsoException;
 use LaravelEnso\FormBuilder\app\Classes\FormBuilder;
 use LaravelEnso\RoAddresses\app\Models\Address;
 use LaravelEnso\RoAddresses\app\Models\County;
+use Symfony\Component\Translation\Exception\InvalidResourceException;
 
 class AddressesController extends Controller
 {
@@ -78,6 +79,8 @@ class AddressesController extends Controller
      */
     public function destroy(Address $address)
     {
+        if($address->is_default) throw new EnsoException(__('The default address cannot be deleted'));
+
         $address->delete();
 
         return [
@@ -140,11 +143,11 @@ class AddressesController extends Controller
      */
     private function getAddressableClass()
     {
-        $class = config('addresses.addressables.'.request()->get('type'));
+        $class = config('enso.addresses.addressables.'.request()->get('type'));
 
         if (!$class) {
             throw new EnsoException(
-                __('Current entity does not exist in contacts.php config file: ').request()->get('type')
+                __('Current entity does not exist in enso/addresses.php config file: ').request()->get('type')
             );
         }
